Create basicPerceptron.cu
in this example the network recognises letters in the neural network theses letters are a,b,c,d, and e. each letter is denoted as an array of 1's and 0's these numbers represent the state of a pixel in an image i chose not to use actual images as i didn't want to overcomplicate the code.